# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: OpenScholar REST API
  version: 0.1.0
  summary: Bounded scholarly discovery, legal-access, library, job, and privacy APIs.
  description: |
    Static contract for every current `/api/v1` endpoint exposed by the Spring Boot
    backend. Research-provider PDF URLs in search results are provider-reported and
    are not independently verified legal-access claims; use the paper access routes
    for verified locations. The API never returns or retains PDF document bytes.

    Security requirements describe hosted mode (`OIDC_SECURITY_ENABLED=true`). In
    loopback local mode, REST requests are permitted without a bearer token and are
    resolved as the fixed local owner. In hosted mode, the configured resource server
    validates JWT signature, expiry, issuer, and audience before applying the scopes
    shown on each operation. Search snapshots, collections, saved papers, privacy
    data, and `SEARCH_METADATA` refresh jobs are resolved for the authenticated
    issuer-and-subject principal; an identifier owned by another principal is returned
    as `404`, not `403`. Canonical papers, stored access records, and `PAPER_ACCESS`
    refresh jobs are shared records as described by their operations.
  license:
    name: All rights reserved
    identifier: LicenseRef-Proprietary
...
servers:
  - url: /
    description: Same-origin deployment
  - url: http://localhost:8080
    description: Loopback backend development
...
paths:
...
  /api/v1/collections:
    get:
      tags: [Collections]
      operationId: listCollections
      summary: List the current principal's collections
      security:
        - HostedOidc: [openscholar.library]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of principal-owned collection summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Collections]
      operationId: createCollection
      summary: Create a collection for the current principal
      security:
        - HostedOidc: [openscholar.library]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionRequest"
      responses:
        "201":
          description: The created collection summary.
          headers:
            Location:
              $ref: "#/components/headers/Location"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/collections/{collectionId}:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    get:
      tags: [Collections]
      operationId: getCollection
      summary: Read one principal-owned collection
      description: Missing and other-principal collection IDs both return `COLLECTION_NOT_FOUND`.
      security:
        - HostedOidc: [openscholar.library]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Collection metadata and a page of saved papers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionDetails"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Collections]
      operationId: updateCollection
      summary: Replace a collection's editable metadata
      description: Both name and description are interpreted as the complete editable state.
      security:
        - HostedOidc: [openscholar.library]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionRequest"
      responses:
        "200":
          description: The updated collection summary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Collections]
      operationId: deleteCollection
      summary: Delete a principal-owned collection
      security:
        - HostedOidc: [openscholar.library]
      responses:
        "204":
          description: The collection and its memberships were deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/collections/{collectionId}/offline-pack:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
    get:
      tags: [Collections]
      operationId: getOfflineCollectionPack
      summary: Read a bounded metadata-only offline snapshot
      description: |
        Returns one complete, replace-in-full snapshot of a principal-owned collection. The
        snapshot is assembled only from stored canonical metadata and never invokes a research
        or access provider. Papers are ordered by `paperId`; author names, identifiers, ISBNs,
        ISSNs, and tags use stable canonical ordering. Missing and other-principal collection IDs
        both return `COLLECTION_NOT_FOUND`.

        Version 1 contains collection name/description, bibliographic metadata, reading status,
        and tags only. It excludes abstracts/full text, PDFs, access/provider/provenance URLs,
        citation or operational metrics, jobs, owner identifiers, and mutation/synchronization
        state. The server never truncates a pack: collections over 500 papers or whose exact
        serialized UTF-8 JSON exceeds 1,048,576 bytes return `OFFLINE_PACK_TOO_LARGE`.
      security:
        - HostedOidc: [openscholar.library]
      responses:
        "200":
          description: The complete metadata-only collection snapshot.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
            X-Content-Type-Options:
              $ref: "#/components/headers/NoSniff"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfflineCollectionPack"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: The collection is absent or not visible to this principal.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "422":
          description: The collection exceeds the item or exact serialized-byte boundary.
          headers:
            Cache-Control:
              $ref: "#/components/headers/NoStore"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
  /api/v1/collections/{collectionId}/papers/{paperId}:
    parameters:
      - $ref: "#/components/parameters/CollectionId"
      - $ref: "#/components/parameters/PaperId"
    put:
      tags: [Collections]
      operationId: putSavedPaper
      summary: Add or replace a saved-paper membership
      description: Idempotently creates or replaces reading status and normalized tags.
      security:
        - HostedOidc: [openscholar.library]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedPaperRequest"
      responses:
        "200":
          description: The created or replaced saved-paper membership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPaper"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Collections]
      operationId: updateSavedPaper
      summary: Replace saved-paper status and tags
      description: Requires an existing membership; the body is the complete mutable state.
      security:
        - HostedOidc: [openscholar.library]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavedPaperRequest"
      responses:
        "200":
          description: The updated saved-paper membership.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedPaper"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Collections]
      operationId: removeSavedPaper
      summary: Remove a paper from a collection
      description: The removal is idempotent once the principal-owned collection has been resolved.
      security:
        - HostedOidc: [openscholar.library]
      responses:
        "204":
          description: The membership is absent.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    HostedOidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example.com/.well-known/openid-configuration
      description: |
        Configurable hosted OIDC/OAuth 2.0 provider. The URL is illustrative; use the
        discovery document for the configured issuer. Access tokens must be JWTs accepted
        by the backend issuer and audience validators. Local loopback mode does not require
        this scheme for REST requests.
  parameters:
...
    PaperId:
      name: paperId
      in: path
      required: true
      description: Shared canonical paper UUID.
      schema:
        $ref: "#/components/schemas/Uuid"
...
    CollectionId:
      name: collectionId
      in: path
      required: true
      description: Principal-owned collection UUID.
      schema:
        $ref: "#/components/schemas/Uuid"
...
    Page:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    PageSize:
      name: size
      in: query
      required: false
      description: Number of items per page.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
...
  headers:
    Location:
      description: Relative canonical URI of the created resource.
      schema:
        type: string
        format: uri-reference
...
    NoSniff:
      description: Prevents browser content-type sniffing.
      schema:
        type: string
        const: nosniff
    NoStore:
      description: Prevents storage of privacy-sensitive or owner-specific responses.
      schema:
        type: string
        const: no-store
...
  responses:
    BadRequest:
      description: Malformed input, unsupported format, or a request-bound validation failure.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: Hosted mode requires a valid bearer access token.
      headers:
        WWW-Authenticate:
          description: Bearer authentication challenge.
          schema:
            type: string
    Forbidden:
      description: The valid hosted token does not grant the operation's required scope.
    NotFound:
      description: The shared resource is absent, or the owner-scoped resource is not visible to this principal.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
  schemas:
...
    Uuid:
      type: string
      format: uuid
    Instant:
      type: string
      format: date-time
    LocalDate:
      type: string
      format: date
    NullableString:
      type: [string, "null"]
    NullableInteger:
      type: [integer, "null"]
      format: int32
...
    NullableLocalDate:
      oneOf:
        - $ref: "#/components/schemas/LocalDate"
        - type: "null"
...
    DocumentType:
      type: string
      enum:
        - ARTICLE
        - PREPRINT
        - CONFERENCE_PAPER
        - THESIS
        - DISSERTATION
        - BOOK
        - BOOK_CHAPTER
        - REPORT
        - DATASET
        - OTHER
    PaperIdentifierType:
      type: string
      enum: [DOI, ARXIV, OPENALEX, PMID, PMCID, CORE, REPOSITORY]
...
    ReadingStatus:
      type: string
      enum: [UNREAD, READING, COMPLETED]
...
    ProblemDetail:
      type: object
      description: RFC 9457 problem detail with stable OpenScholar extension fields.
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri
          examples: ["urn:openscholar:problem:validation-failed"]
        title:
          type: string
        status:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          description: Stable machine-readable error code.
          enum:
            - VALIDATION_FAILED
            - INVALID_REQUEST
            - INVALID_PAPER_IDENTIFIER
            - UNSUPPORTED_CITATION_FORMAT
            - SEARCH_NOT_FOUND
            - SEARCH_PAGE_EXHAUSTED
            - PAPER_NOT_FOUND
            - PAPER_IDENTIFIER_NOT_FOUND
            - COLLECTION_NOT_FOUND
            - OFFLINE_PACK_TOO_LARGE
            - SAVED_PAPER_NOT_FOUND
            - REFRESH_JOB_NOT_FOUND
            - REFRESH_JOB_NOT_RETRYABLE
            - SEARCH_PROVIDER_UNAVAILABLE
            - SEARCH_COORDINATION_TIMEOUT
            - SEARCH_COORDINATION_INTERRUPTED
            - SEARCH_DEADLINE_EXCEEDED
            - SEARCH_EXECUTION_INTERRUPTED
            - ACCESS_PROVIDERS_UNAVAILABLE
            - ACCESS_REFRESH_RATE_LIMITED
        violations:
          type: array
          items:
            $ref: "#/components/schemas/ValidationViolation"
        retryable:
          type: boolean
        retryAfterSeconds:
          type: integer
          format: int64
          minimum: 1
      additionalProperties: true
    ValidationViolation:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
      additionalProperties: false
...
    PaperIdentifier:
      type: object
      required: [type, namespace, value]
      properties:
        type:
          $ref: "#/components/schemas/PaperIdentifierType"
        namespace:
          type: string
        value:
          type: string
      additionalProperties: false
...
    CollectionRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          pattern: '\S'
          description: Whitespace is collapsed before persistence.
        description:
          type: [string, "null"]
          maxLength: 1000
          description: Whitespace is collapsed; blank becomes null.
      additionalProperties: false
    SavedPaperRequest:
      type: object
      required: [readingStatus, tags]
      properties:
        readingStatus:
          $ref: "#/components/schemas/ReadingStatus"
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 40
            pattern: '\S'
          description: Tags are whitespace-collapsed and lowercased; normalized duplicates are rejected and responses are sorted.
      additionalProperties: false
    OfflineCollectionPack:
      type: object
      description: |
        A manual, replace-in-full version 1 snapshot. It is not a synchronization cursor and
        contains no document bytes, access locations, provider URLs, or operational data.
      required: [schemaVersion, generatedAt, collection, papers]
      properties:
        schemaVersion:
          type: integer
          format: int32
          const: 1
        generatedAt:
          $ref: "#/components/schemas/Instant"
        collection:
          $ref: "#/components/schemas/OfflineCollectionMetadata"
        papers:
          type: array
          maxItems: 500
          items:
            $ref: "#/components/schemas/OfflineCollectionPaper"
      additionalProperties: false
    OfflineCollectionMetadata:
      type: object
      required: [collectionId, name, description]
      properties:
        collectionId:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, "null"]
          maxLength: 1000
      additionalProperties: false
    OfflineCollectionPaper:
      type: object
      description: Stored bibliographic metadata plus this collection's reading state and tags.
      required:
        - paperId
        - title
        - authors
        - publicationDate
        - publicationYear
        - documentType
        - language
        - venueName
        - identifiers
        - publisher
        - institution
        - volume
        - issue
        - pages
        - articleNumber
        - edition
        - isbn
        - issn
        - degree
        - readingStatus
        - tags
      properties:
        paperId:
          $ref: "#/components/schemas/Uuid"
        title:
          type: string
          minLength: 1
        authors:
          type: array
          description: Credited display names in author-position order; external author IDs are omitted.
          items:
            type: string
        publicationDate:
          $ref: "#/components/schemas/NullableLocalDate"
        publicationYear:
          $ref: "#/components/schemas/NullableInteger"
        documentType:
          $ref: "#/components/schemas/DocumentType"
        language:
          $ref: "#/components/schemas/NullableString"
        venueName:
          $ref: "#/components/schemas/NullableString"
        identifiers:
          type: array
          description: Stored bibliographic identifier values; never access or provider URLs.
          items:
            $ref: "#/components/schemas/PaperIdentifier"
        publisher:
          $ref: "#/components/schemas/NullableString"
        institution:
          $ref: "#/components/schemas/NullableString"
        volume:
          $ref: "#/components/schemas/NullableString"
        issue:
          $ref: "#/components/schemas/NullableString"
        pages:
          $ref: "#/components/schemas/NullableString"
        articleNumber:
          $ref: "#/components/schemas/NullableString"
        edition:
          $ref: "#/components/schemas/NullableString"
        isbn:
          type: array
          items:
            type: string
        issn:
          type: array
          items:
            type: string
        degree:
          $ref: "#/components/schemas/NullableString"
        readingStatus:
          $ref: "#/components/schemas/ReadingStatus"
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 40
      additionalProperties: false
    CollectionSummary:
      type: object
      required: [collectionId, name, description, paperCount, createdAt, updatedAt]
      properties:
        collectionId:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, "null"]
          maxLength: 1000
        paperCount:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
      additionalProperties: false
    CollectionDetails:
      type: object
      required: [collectionId, name, description, paperCount, createdAt, updatedAt, papers]
      properties:
        collectionId:
          $ref: "#/components/schemas/Uuid"
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, "null"]
          maxLength: 1000
        paperCount:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
        papers:
          $ref: "#/components/schemas/SavedPaperPage"
      additionalProperties: false
    SavedPaper:
      type: object
      required:
        - collectionId
        - collectionName
        - paperId
        - title
        - authors
        - publicationYear
        - documentType
        - readingStatus
        - tags
        - savedAt
        - updatedAt
      properties:
        collectionId:
          $ref: "#/components/schemas/Uuid"
        collectionName:
          type: string
        paperId:
          $ref: "#/components/schemas/Uuid"
        title:
          type: string
        authors:
          type: array
          items:
            type: string
        publicationYear:
          $ref: "#/components/schemas/NullableInteger"
        documentType:
          $ref: "#/components/schemas/DocumentType"
        readingStatus:
          $ref: "#/components/schemas/ReadingStatus"
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 40
        savedAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
      additionalProperties: false
    CollectionPage:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/CollectionSummary"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
      additionalProperties: false
    SavedPaperPage:
      type: object
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/SavedPaper"
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
      additionalProperties: false
...